pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "catalystindex"
version = "0.1.0"
description = "Catalyst Index RAG platform core services"
readme = "README.md"
authors = [{name = "Catalyst Index"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi<0.115",
    "uvicorn",
    "pydantic>=2.0.0",
    "pydantic-settings",
    "redis>=5.0.0",
    "rq>=1.15.1",
    "psycopg[binary]>=3.2.0",
    "qdrant-client>=1.15.1",
    "openai>=1.30.0",
    "tiktoken>=0.8.0",
    "boto3>=1.35.0",
    "cohere>=5.0.0",
    "pdfminer-six>=20221105",
    "pillow>=11.0.0",
    "pypdf>=5.0.0",
    "python-magic>=0.4.27",
    "pi-heif>=0.20.0",
    "pikepdf>=9.0.0",
    "pymupdf>=1.25.0",
    "unstructured[pdf,pptx,docx,xlsx,image]>=0.16.0",
    "openparse>=0.7.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0"
]
qdrant = [
    "qdrant-client>=1.15.1",
]
redis = [
    "redis>=5.0.0",
]
openai = [
    "openai>=1.13.3",
    "tiktoken>=0.5.0",
]
workers = [
    "rq>=1.15.1",
    "redis>=5.0.0",
    "psycopg[binary]>=3.1.12",
]
ingestion = [
    "pdfminer-six>=20221105",
    "pillow>=11.0.0",
    "pypdf>=5.0.0",
    "python-magic>=0.4.27",
    "pi-heif>=0.20.0",
    "pikepdf>=9.0.0",
    "pymupdf>=1.25.0",
    "unstructured[pdf,pptx,docx,xlsx,image]>=0.16.0",
    "openparse>=0.7.0",
    "openai>=1.30.0",
    "tiktoken>=0.8.0",
    "scikit-learn>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests that require external services (Qdrant, etc.)",
]
